requirements.txt
anyio==3.6.2
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
ecdsa==0.18.0
fastapi==0.95.1
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
mysql-connector-python==8.0.33
passlib==1.7.4
protobuf==3.20.3
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.7
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.12
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0.3